refactor(FE): `useIsRegistered` hook
