+++ frontend/src/components/Battle/Battle.js
@@ -3,12 +3,12 @@ import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
 import { useParams } from "react-router-dom";
-import { useIsRegistered } from "../../hooks/queries";
 import useAuth from "../../hooks/useAuth";
 import useCurrentBattle from "../../hooks/useCurrentBattle";
 import useFakemonsByGym from "../../hooks/useFakemonsByGym";
 import useFakemonsByUser from "../../hooks/useFakemonsByUser";
 import useGyms from "../../hooks/useGyms";
+import useIsRegistered from "../../hooks/useIsRegistered";
 import Alert from "../Alert";
 import ImportWalletAlert from "../ImportWalletAlert";
 import Actions from "./Actions";
