+++ frontend/src/components/Profile/Profile.js
@@ -2,7 +2,7 @@ import React from "react";
 import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
 import { Link } from "react-router-dom";
-import { useIsRegistered, useTokenBalance } from "../../hooks/queries";
+import { useTokenBalance } from "../../hooks/queries";
 import ImportWalletAlert from "../ImportWalletAlert";
 import Alert from "../Alert";
 import { useMintFakemon, useRegister } from "../../hooks/mutations";
@@ -11,6 +11,7 @@ import ButtonWithLoader from "../ButtonWithLoader";
 import useAuth from "../../hooks/useAuth";
 import useToast from "../../hooks/useToast";
 import useFakemonsByUser from "../../hooks/useFakemonsByUser";
+import useIsRegistered from "../../hooks/useIsRegistered";
 
 function Profile() {
   const { walletAddress } = useAuth();
