+++ frontend/src/hooks/queries.js
@@ -70,19 +70,3 @@ export function useTokenBalance() {
     }
   );
 }
-
-export function useIsRegistered() {
-  const { data: blockchain } = useBlockchain();
-
-  return useQuery(
-    [QueryKeys.IS_REGISTERED],
-    () => {
-      try {
-        return blockchain.fakemon.users(blockchain.signerAddress);
-      } catch (error) {
-        return false;
-      }
-    },
-    { enabled: !!blockchain, placeholderData: false }
-  );
-}
